feat(mhxy): 账号增加封禁状态
