+++ src/entities/mhxy/mhxy-account.entity.ts
@@ -1,15 +1,15 @@
 import { Column, Entity, JoinColumn, ManyToOne, OneToOne, PrimaryColumn } from 'typeorm';
 import { TimeNotSelectBase } from '../base';
-import { MHXY_ACCOUNT_LENGTH } from '../../constants';
+import { MHXY_ACCOUNT_KEY_LENGTH, ENUM_MHXY_ACCOUNT_STATUS } from '../../constants';
 import { MhxyAccountGroup, MhxyAccountGroupItem, User } from '../../entities';
 
 /** 梦幻账号表 */
-@Entity({ name: 'mhxy_account', orderBy: { isPrimary: 'DESC' } })
+@Entity({ name: 'mhxy_account', orderBy: { isPrimary: 'DESC', status: 'ASC' } })
 export class MhxyAccount extends TimeNotSelectBase {
-  @PrimaryColumn({ length: MHXY_ACCOUNT_LENGTH.ID_MAX, comment: '账号 Id （官方）' })
+  @PrimaryColumn({ length: MHXY_ACCOUNT_KEY_LENGTH.ID_MAX, comment: '账号 Id （官方）' })
   id: string;
 
-  @Column({ length: MHXY_ACCOUNT_LENGTH.NAME_MAX, comment: '账号名称' })
+  @Column({ length: MHXY_ACCOUNT_KEY_LENGTH.NAME_MAX, comment: '账号名称' })
   name: string;
 
   @Column({ name: 'is_primary', comment: '是否是主号', default: false })
@@ -26,6 +26,14 @@ export class MhxyAccount extends TimeNotSelectBase {
 
   @Column({ name: 'lock_gold', comment: '被锁金币数量', default: 0 })
   lockGold: number;
+
+  @Column({
+    comment: '状态',
+    type: 'enum',
+    enum: ENUM_MHXY_ACCOUNT_STATUS,
+    default: ENUM_MHXY_ACCOUNT_STATUS.ACTIVE,
+  })
+  status: ENUM_MHXY_ACCOUNT_STATUS;
 
   /** 归属系统用户 */
   @ManyToOne(() => User, {
