+++ src/mhxy/dto/index.ts
@@ -2,74 +2,11 @@ export * from './channel.dto';
 export * from './gold-record.dto';
 export * from './gold-transfer.dto';
 export * from './account-group.dto';
+export * from './account.dto';
 
-import { Allow, IsBoolean, IsInt, IsNotEmpty, Length, Min, Validate } from 'class-validator';
-import { MHXY_ACCOUNT_LENGTH, MHXY_PROP_CATEGORY_LENGTH } from 'src/constants';
-import { MhxyRoleValidator, MhxySectValidator } from './custom-validation';
+import { Allow, IsBoolean, IsNotEmpty, Length } from 'class-validator';
+import { MHXY_PROP_CATEGORY_LENGTH } from 'src/constants';
 import { MhxyPropCategory } from 'src/entities';
-
-export class MhxyAccountIdDto {
-  @IsNotEmpty({ message: 'id 不能为空' })
-  @Length(1, MHXY_ACCOUNT_LENGTH.ID_MAX, {
-    message: `id 长度为 1 - ${MHXY_ACCOUNT_LENGTH.ID_MAX} 个字符`,
-  })
-  /** 账号 Id （官方） */
-  id: string;
-}
-
-/** 账号编辑 */
-export class MhxyAccountEditDto extends MhxyAccountIdDto {
-  @IsNotEmpty({ message: 'name 不能为空' })
-  @Length(1, MHXY_ACCOUNT_LENGTH.NAME_MAX, {
-    message: `name 长度为 1 - ${MHXY_ACCOUNT_LENGTH.NAME_MAX} 个字符`,
-  })
-  /** 账号名称 */
-  name: string;
-
-  @IsNotEmpty({ message: 'role 不能为空' })
-  @Validate(MhxyRoleValidator, {
-    message: `未知的 role`,
-  })
-  /** 角色 */
-  role: string;
-
-  @IsNotEmpty({ message: 'sect 不能为空' })
-  @Validate(MhxySectValidator, {
-    message: `未知的 sect`,
-  })
-  /** 门派 */
-  sect: string;
-  @Allow()
-  /** 是否是主号 */
-  isPrimary: boolean;
-  @Allow()
-  /** 分组 id */
-  groupId: number;
-  @Allow()
-  /** 分组备注  */
-  groupRemark: string;
-}
-
-/** 账号 */
-export class MhxyAccountDto extends MhxyAccountEditDto {
-  @IsNotEmpty({ message: 'isPrimary 不能为空' })
-  /** 是否是主号 */
-  isPrimary: boolean;
-  @IsInt({ message: 'gold 必须为整数' })
-  @Min(0, { message: 'gold 必须为大于 0 的数字' })
-  /**  金币数量 */
-  gold: number;
-  @IsInt({ message: 'lockGold 必须为整数' })
-  /**  金币数量 */
-  lockGold: number;
-}
-
-/** 贸易种类查询 */
-export class GoldTradeCategorySearchDto {
-  /** 是否是转金项 */
-  @Allow()
-  isTransfer?: boolean;
-}
 
 /** 贸易种类新增 */
 export class GoldTradeCategoryAddDto {
